gcp/requirements.txt
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
fp-NGFW-SMC-python==1.0.16
google-api-core==2.8.1
google-auth==2.7.0
google-cloud-compute==1.3.2
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
idna==3.3
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.20.0
pytz==2021.1
requests==2.28.0
rsa==4.8
six==1.16.0
urllib3==1.26.9
